fix bugs

+++ blogicum/blog/views.py
@@ -1,62 +1,55 @@
+
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.shortcuts import get_object_or_404, redirect, render
 
 from blog.forms import CommentForm, PostForm, ProfileForm
-from blog.models import Comment, Post
-from blog.utils import posts_pagination, query_category, query_post
-
-
-# from django.contrib.auth.mixins import LoginRequiredMixin
-# from django.views.generic import UpdateView
-# from django.urls import reverse, reverse_lazy
-
+from blog.models import Category, Comment, Post
+from blog.utils import posts_pagination, query_post
 
 
 def index(request):
 
     page_obj = posts_pagination(request, query_post(
         filters=True,
-        comments=True
+        with_comments=True
     ))
-    context = {'page_obj':page_obj}
+    context = {'page_obj': page_obj}
     return render(request, 'blog/index.html', context)
 
 
 def category_posts(request, category_slug):
 
     category = get_object_or_404(
-        query_category(),
+        Category,
         slug=category_slug,
+        is_published=True,
     )
     page_obj = posts_pagination(
         request,
-        query_post(user=category.posts)
+        query_post(manager=category.posts)
     )
     context = {'category': category, 'page_obj': page_obj}
     return render(request, 'blog/category.html', context)
 
 
-def post_detail(request, id):
+def post_detail(request, post_id):
 
-    post = get_object_or_404(
-        Post,
-        pk=id,
-    )
+    post = get_object_or_404(Post, id=post_id)
     if post.author != request.user:
-        post = get_object_or_404(query_post(), id=id)
-    comments = Comment.objects.select_related('author').filter(post=post)
+        post = get_object_or_404(query_post(), id=post_id)
+    comments = post.comments.order_by('created_at')
     form = CommentForm()
     context = {
         'post': post,
         'form': form,
-        'comments': comments,
+        'comments': comments
     }
     return render(request, 'blog/detail.html', context)
 
 
 @login_required
-def create_posts(request):
+def create_post(request):
 
     form = PostForm(request.POST or None, files=request.FILES or None)
     if form.is_valid():
@@ -69,29 +62,31 @@ def create_posts(request):
 
 
 @login_required
-def edit_posts(request, id_post):
+def edit_post(request, post_id):
 
-    post = get_object_or_404(Post, id=id_post)
+    post = get_object_or_404(Post, id=post_id)
     if request.user != post.author:
-        return redirect('blog:post_detail', id_post)
+        return redirect('blog:post_detail', post_id)
     form = PostForm(request.POST or None, instance=post)
     if form.is_valid():
         form.save()
-        return redirect('blog:post_detail', id_post)
+        return redirect('blog:post_detail', post_id)
     context = {'form': form}
     return render(request, 'blog/create.html', context)
 
 
 @login_required
-def delete_posts(request, id_post):
+def delete_post(request, post_id):
 
-    post = get_object_or_404(Post, id=id_post)
+    post = get_object_or_404(Post, id=post_id)
     if request.user != post.author:
-        return redirect('blog:post_detail', id_post)
-    form = PostForm(request.POST or None, instance=post)
+        return redirect('blog:post_detail', post_id)
     if request.method == 'POST':
+        form = PostForm(request.POST or None, instance=post)
         post.delete()
         return redirect('blog:index')
+    else:
+        form = PostForm(instance=post)
     context = {'form': form}
     return render(request, 'blog/create.html', context)
 
@@ -99,65 +94,70 @@ def delete_posts(request, id_post):
 def profile(request, username):
 
     profile = get_object_or_404(User, username=username)
-    posts = query_post(user=profile.posts) if profile != request.user else query_post(user=profile.posts, filters=False)
-
+    if profile == request.user:
+        posts = query_post(manager=profile.posts,
+                           filters=False,)
+    else:
+        posts = query_post(manager=profile.posts,)
     page_obj = posts_pagination(request, posts)
-    context = {
-        'profile': profile,
-        'page_obj': page_obj,
-    }
-
+    context = {'profile': profile,
+               'page_obj': page_obj}
     return render(request, 'blog/profile.html', context)
 
 
 @login_required
 def edit_profile(request):
 
-    form = ProfileForm(request.POST or None, instance=request.user)
-    if form.is_valid():
-        form.save()
-        return redirect('blog:profile', username=request.user.username)
-    return render(request, 'blog/user.html', {'form': form})
+    if request.method == 'POST':
+        form = ProfileForm(request.POST, instance=request.user)
+        if form.is_valid():
+            form.save()
+            return redirect('blog:profile', request.user)
+    else:
+        form = ProfileForm(instance=request.user)
+    context = {'form': form}
+    return render(request, 'blog/user.html', context)
 
 
 @login_required
-def add_comment(request, id_post):
-    """Добавление комментария публикации."""
-    post = get_object_or_404(Post, id=id_post)
+def add_comment(request, post_id):
+
+    post = get_object_or_404(Post, id=post_id)
     form = CommentForm(request.POST or None)
     if form.is_valid():
         comment = form.save(commit=False)
-        comment.author = request.user
         comment.post = post
+        comment.author = request.user
         comment.save()
-    return redirect('blog:post_detail', id=id_post)
+    return redirect('blog:post_detail', post_id)
 
 
 
 @login_required
-def edit_comment(request, id_post, id_comment):
+def edit_comment(request, post_id, comment_id):
 
-    comment = get_object_or_404(Comment,
-                                id=id_comment,
-                                post_id=id_post,
-                                author=request.user)
-    form = CommentForm(request.POST or None, instance=comment)
+    comment = get_object_or_404(Comment, id=comment_id)
+    if request.user != comment.author:
+        return redirect('blog:post_detail', post_id)
+    if request.method == "POST":
+        form = CommentForm(request.POST or None, instance=comment)
+        if form.is_valid():
+            form.save()
+            return redirect('blog:post_detail', post_id)
+    else:
+        form = CommentForm(instance=comment)
     context = {'form': form, 'comment': comment}
-    if form.is_valid():
-        form.save()
-        return redirect('blog:post_detail', id=id_post)
     return render(request, 'blog/comment.html', context)
 
 
 @login_required
-def delete_comment(request, id_post, id_comment):
+def delete_comment(request, post_id, comment_id):
 
-    comment = get_object_or_404(Comment,
-                                id=id_comment,
-                                post_id=id_post,
-                                author=request.user)
-    context = {'comment': comment}
-    if request.method == 'POST':
+    comment = get_object_or_404(Comment, id=comment_id)
+    if request.user != comment.author:
+        return redirect('blog:post_detail', post_id)
+    if request.method == "POST":
         comment.delete()
-        return redirect('blog:post_detail', id=id_post)
+        return redirect('blog:post_detail', post_id)
+    context = {'comment': comment}
     return render(request, 'blog/comment.html', context)
